backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
requests==2.32.3
pydantic==2.7.1
python-dotenv==1.0.1
